Additions  - You can now customize what effects the player has while downed.  - You can now disable movement and jumping  - You can disable the give up option
 Fixes
 - Death event has been moved to high priority to avoid compatibility issues

+++ src/main/java/invoker54/reviveme/common/capability/FallenCapability.java
@@ -137,7 +137,7 @@ public class FallenCapability {
 //        if (ReviveMeConfig.timeLeft == 0) return 1;
 
         if (divideByMax)
-            return 1 - ((level.getGameTime() - fellStart)/(float) fellEnd);
+            return 1 - ((level.getGameTime() - fellStart)/ fellEnd);
 
         return ((fellStart + fellEnd) - level.getGameTime())/20f;
     }
